mailchimprest/requirements.txt
amqp==2.5.1
appnope==0.1.0
awsebcli==3.15.3
backcall==0.1.0
billiard==3.6.1.0
blessed==1.15.0
boto3==1.9.215
botocore==1.12.215
cached-property==1.5.1
celery==4.3.0
cement==2.8.2
certifi==2019.6.16
chardet==3.0.4
colorama==0.3.9
decorator==4.4.0
Django==2.2.1
django-extensions==2.2.1
django-filter==2.2.0
djangorestframework==3.10.2
docker==3.7.2
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
future==0.16.0
idna==2.7
importlib-metadata==0.19
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.14.1
jmespath==0.9.4
jsonschema==2.6.0
kombu==4.6.4
mailchimp3==3.0.7
Markdown==3.1.1
more-itertools==7.2.0
parso==0.5.1
pathspec==0.5.9
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycurl==7.43.0.3
Pygments==2.4.2
python-dateutil==2.8.0
pytz==2019.1
PyYAML==3.13
requests==2.20.1
s3transfer==0.2.1
semantic-version==2.5.0
six==1.11.0
sqlparse==0.3.0
termcolor==1.1.0
texttable==0.9.1
traitlets==4.3.2
urllib3==1.24.3
vine==1.3.0
wcwidth==0.1.7
websocket-client==0.56.0
zipp==0.6.0
